refactor(agent-service): modify the architecture of the service
- add database module to the infrastructure
