+++ apps/agent/src/application/commands/handlers/update-refresh-token.handler.ts
@@ -1,19 +1,27 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { UpdateRefreshTokenCommand } from '../impl/update-refresh-token.command';
-import { AgentEntityRepository } from '../../../domain/base-agent.entity-repo';
+import { Inject } from '@nestjs/common';
+import {
+  AGENT_PROVIDER,
+  IAgentProvider,
+} from '../../../infrastructure/database/providers/agent.provider';
 
 @CommandHandler(UpdateRefreshTokenCommand)
 export class UpdateRefreshTokenHandler
   implements ICommandHandler<UpdateRefreshTokenCommand, void>
 {
-  constructor(private readonly agnetRepo: AgentEntityRepository) {}
+  z;
+
+  constructor(
+    @Inject(AGENT_PROVIDER) private readonly agentProvider: IAgentProvider,
+  ) {}
 
   async execute(command: UpdateRefreshTokenCommand): Promise<void> {
-    const agent = await this.agnetRepo.findById(command.agentId);
+    const agent = await this.agentProvider.findById(command.agentId);
     if (!agent) return;
 
     agent.changeRefreshToken(command.refreshToken);
-    await this.agnetRepo.save(agent);
+    await this.agentProvider.save(agent);
     agent.commit();
   }
 }
