+++ apps/agent/src/application/queries/handlers/get-account-agents-ids.handler.ts
@@ -1,14 +1,20 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { GetAccountAgentsIdsQuery } from '../impl/get-account-agents-ids.query';
-import { AgentQueryRepository } from '../../../infrastructure/repositories/agent.query-repo';
+import { Inject } from '@nestjs/common';
+import {
+  AGENT_PROVIDER,
+  IAgentProvider,
+} from '../../../infrastructure/database/providers/agent.provider';
 
 @QueryHandler(GetAccountAgentsIdsQuery)
 export class GetAccountAgentsIdsHandler
   implements IQueryHandler<GetAccountAgentsIdsQuery, string[]>
 {
-  constructor(private readonly agentRepo: AgentQueryRepository) {}
+  constructor(
+    @Inject(AGENT_PROVIDER) private readonly agentProvider: IAgentProvider,
+  ) {}
 
   async execute(query: GetAccountAgentsIdsQuery): Promise<string[]> {
-    return await this.agentRepo.findIdsByAccount(query.accountId);
+    return await this.agentProvider.findIdsByAccount(query.accountId);
   }
 }
