+++ apps/agent/src/application/queries/handlers/get-account-agents.handler.ts
@@ -1,15 +1,21 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
-import { AgentModel } from '../../../infrastructure/models/agent.model';
-import { AgentQueryRepository } from '../../../infrastructure/repositories/agent.query-repo';
 import { GetAccountAgentsQuery } from '../impl/get-account-agents.query';
+import { Inject } from '@nestjs/common';
+import {
+  AGENT_PROVIDER,
+  IAgentProvider,
+} from '../../../infrastructure/database/providers/agent.provider';
+import { AgentModel } from '../../../infrastructure/database/mongo/models/agent.model';
 
 @QueryHandler(GetAccountAgentsQuery)
 export class GetAccountAgentsHandler
   implements IQueryHandler<GetAccountAgentsQuery, AgentModel[]>
 {
-  constructor(private readonly agentRepo: AgentQueryRepository) {}
+  constructor(
+    @Inject(AGENT_PROVIDER) private readonly agentProvider: IAgentProvider,
+  ) {}
 
   async execute(query: GetAccountAgentsQuery): Promise<AgentModel[]> {
-    return await this.agentRepo.findByAccount(query.accountId);
+    return await this.agentProvider.findByAccount(query.accountId);
   }
 }
