+++ apps/agent/src/application/queries/handlers/get-by-email.handler.ts
@@ -1,15 +1,21 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { GetByEmailQuery } from '../impl/get-by-email.query';
-import { AgentQueryRepository } from '../../../infrastructure/repositories/agent.query-repo';
-import { AgentModel } from '../../../infrastructure/models/agent.model';
+import { Inject } from '@nestjs/common';
+import {
+  AGENT_PROVIDER,
+  IAgentProvider,
+} from '../../../infrastructure/database/providers/agent.provider';
+import { AgentModel } from '../../../infrastructure/database/mongo/models/agent.model';
 
 @QueryHandler(GetByEmailQuery)
 export class GetByEmailHandler
   implements IQueryHandler<GetByEmailQuery, AgentModel | null>
 {
-  constructor(private readonly agentRepo: AgentQueryRepository) {}
+  constructor(
+    @Inject(AGENT_PROVIDER) private readonly agentProvider: IAgentProvider,
+  ) {}
 
   async execute({ email }: GetByEmailQuery): Promise<AgentModel | null> {
-    return this.agentRepo.findByEmail(email);
+    return this.agentProvider.findByEmail(email);
   }
 }
