+++ apps/agent/src/application/queries/handlers/get-by-id.handler.ts
@@ -1,15 +1,19 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { GetByIdQuery } from '../impl/get-by-id.query';
-import { AgentQueryRepository } from '../../../infrastructure/repositories/agent.query-repo';
-import { AgentModel } from '../../../infrastructure/models/agent.model';
+import { Inject } from '@nestjs/common';
+import {
+  AGENT_PROVIDER,
+  IAgentProvider,
+} from '../../../infrastructure/database/providers/agent.provider';
+import { Agent } from '../../../domain/entities/agent.entity';
 
 @QueryHandler(GetByIdQuery)
-export class GetByIdHandler
-  implements IQueryHandler<GetByIdQuery, AgentModel | null>
-{
-  constructor(private readonly agentRepo: AgentQueryRepository) {}
+export class GetByIdHandler implements IQueryHandler<GetByIdQuery, Agent> {
+  constructor(
+    @Inject(AGENT_PROVIDER) private readonly agentProvider: IAgentProvider,
+  ) {}
 
-  async execute({ agentId }: GetByIdQuery): Promise<AgentModel | null> {
-    return this.agentRepo.findById(agentId);
+  async execute({ agentId }: GetByIdQuery): Promise<Agent> {
+    return this.agentProvider.findById(agentId);
   }
 }
